Added skip and reverse cards
Added card to skip and reverse game.
Game is now saving all that data as a big JSON, for easily create a reader to continue game where it was left.

Added functionality for admin when clicking to start new game and user count is less than 2 don't start.

Now when a game finishes you can start one up right after.

Now there is a data attribute to indicate when it's you turn, the browser window will have a white border